on screen state change + constant rate social timer
1. Screen state change fires reciever.
2. Screen ON - new Social timer starts
3. Screen OFF - social timer off

Scoial timer run every 5 sec. When started begin to check state of
social networks every 5 secs, then interval smoothly changes to 60 sec.
Every ten minutes another timer checks where is user and is there are
friends (or nonfriends if it is not enough friends around) near. A that
moment also the social data gets to server.